Backport grok parsing patterns for gorouter logs
We currently have a fork of the upstream repo in order to include a
fix[1]. This fix has now been merged, however, it is only available in
the latest release (v206.0.0) and this in turn requires the latest
version of logsearch which requires a major version upgrade of ES and
Kibana.

We have elected to backport the changes we require in order to avoid
having to do the ES/Kibana upgrade.

[1] https://github.com/cloudfoundry-community/logsearch-for-cloudfoundry/pull/255