Handle RFC 5424 syslog format for CF platform logs
The latest syslog-release project introduced a new Syslog log format which is
based on RFC 5424 [1]. This format is automatically parsed by the Logstash
filters defined in [2] therefore the grok filters will fail in platform.conf.

We check for syslog_sd_id first and make sure the common fields and tags are set.

[1] https://github.com/cloudfoundry/syslog-release/blob/e58e88281429f0bac1313994db447fce22b29106/jobs/syslog_forwarder/templates/rsyslog.conf.erb#L54
[2] https://github.com/cloudfoundry-community/logsearch-boshrelease/blob/9c0376bfd403cc63df058a30e6e27e4f0b41c889/src/logsearch-config/src/logstash-filters/snippets/syslog_standard.conf#L3